Removed nodes from the project
